+++ index.html
@@ -2,19 +2,19 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/logo.png" />
+    <link rel="icon" href="public/image/Logo.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta
       name="대현닷컴"
       content="마피아42의 편리기능 사이트 대현닷컴입니다."
     />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/image/logo.png" />
+    <link rel="apple-touch-icon" href="public/image/Logo.svg" />
 
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    <link rel="manifest" href="manifest.json" />
 
     <meta name="description" content="마피아42의 편리기능 사이트 대현닷컴입니다.">
-    <link rel="canonical" href="../">
+    <link rel="canonical" href="">
     <meta name="robots" content="index,follow">
     <meta property="og:description" content="마피아42의 편리기능 사이트 대현닷컴입니다.">
     <meta property="og:url" content="http://대현.com">
@@ -28,8 +28,8 @@
     <meta name="twitter:image" content="../image/bg.png">
     <meta name="twitter:card" content="summary_large_image">
     
-    <link rel="image_src" href="../image/logo.png">
-    <link rel="shortcut icon" href="../image/logo.png">
+    <link rel="image_src" href="public/image/Logo.svg">
+    <link rel="shortcut icon" href="public/image/Logo.svg">
     <title>대현닷컴</title>
     <!-- 이곳에 위의 Google 태그를 삽입해준다 -->
     <!-- Google tag (gtag.js) -->
@@ -46,6 +46,6 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-
+    <script type="module" src="/src/app/main.jsx"></script>
   </body>
 </html>
