+++ src/components/base/Layout.jsx
@@ -1,0 +1,9 @@
+import styles from './styles/Layout.module.css'
+
+function Layout({children}) {
+    return <div className={styles.Layout}>
+        {children}
+    </div>
+}
+
+export default Layout
