+++ src/components/base/styles/Layout.module.css
@@ -1,0 +1,8 @@
+.Layout {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    height: 100vh;
+    width: 100%;
+}
