+++ src/components/styles/Header.module.css
@@ -45,28 +45,6 @@
   gap: 20px; /* Add space between the a tags (adjust the value as needed) */
 }
 
-.notactive {
-  color: white;
-  margin-top: auto;
-  margin-bottom: 0px;
-  width: 100px;
-  height: 50px;
-  text-align: center;
-  line-height: 50px;
-}
-
-.active {
-  color: white;
-  margin-top: auto;
-  margin-bottom: 0px;
-  width: 100px;
-  height: 50px;
-  text-align: center;
-  line-height: 50px;
-
-  border-bottom: solid 3px white;
-}
-
 .menubar::-webkit-scrollbar {
   height: 8px; /* 세로 스크롤바 높이 */
 }
