+++ src/pages/Board.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./Board.module.css";
+import styles from "./styles/Board.module.css";
 import { useNavigate } from "react-router-dom";
 
 function Board() {
