+++ src/pages/styles/BoardDetail.module.css
