+++ vite.config.js
@@ -1,0 +1,8 @@
+import {defineConfig} from 'vite'
+import react from '@vitejs/plugin-react'
+import svgr from "vite-plugin-svgr";
+
+// https://vite.dev/config/
+export default defineConfig({
+    plugins: [react(), svgr()],
+})
